backend/pyproject.toml
[project]
name = "vanilla-rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "chromadb>=0.6.0",
    "faiss-cpu>=1.9.0",
    "fastapi[standard]>=0.115.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "pillow>=11.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "sentence-transformers>=3.2.0",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]
